chore : 프로젝트 이름 변경
