+++ chat-service/build.gradle
