+++ chat-service/src/main/java/com/example/chatservice/ChatServiceApplication.java
@@ -1,0 +1,19 @@
+package com.example.chatservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@EnableCaching
+@EnableFeignClients
+@EnableEurekaClient
+@SpringBootApplication
+public class ChatServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ChatServiceApplication.class, args);
+	}
+
+}
