+++ chat-service/src/main/java/com/example/chatservice/model/entity/ChatRoom.java
@@ -1,0 +1,24 @@
+package com.example.chatservice.model.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ChatRoom {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "CHAT_ROOM_ID")
+    private Integer id;
+
+    @Column(nullable = false, length = 45)
+    private String name;
+}
