+++ chat-service/src/main/java/com/example/chatservice/repository/ChatRoomRepository.java
@@ -1,0 +1,7 @@
+package com.example.chatservice.repository;
+
+import com.example.chatservice.model.entity.ChatRoom;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
+}
