+++ chat-service/src/main/java/com/example/chatservice/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.chatservice.repository;
+
+import com.example.chatservice.model.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, String> {
+    Optional<User> findByNickname(String nickname);
+}
