+++ chat-service/src/main/java/com/example/chatservice/service/redis/RedisSubscriber.java
@@ -1,0 +1,35 @@
+package com.example.chatservice.service.redis;
+
+import com.example.chatservice.model.dto.ChatPostDto.ChatPostReq;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.redis.connection.Message;
+import org.springframework.data.redis.connection.MessageListener;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.messaging.simp.SimpMessageSendingOperations;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RedisSubscriber implements MessageListener {
+
+    private final ObjectMapper objectMapper;
+    private final RedisTemplate redisTemplate;
+    private final SimpMessageSendingOperations messagingTemplate;
+
+    @Override
+    public void onMessage(Message message, byte[] pattern) {
+
+        try {
+            String publishMessage = (String) redisTemplate.getStringSerializer().deserialize(message.getBody());
+            ChatPostReq chatMessage = objectMapper.readValue(publishMessage, ChatPostReq.class);
+
+            messagingTemplate.convertAndSend("/sub/chat/room/" + chatMessage.getRoomId(), chatMessage);
+        }
+        catch (Exception e) {
+            /* 나중에
+
+             */
+        }
+    }
+}
