+++ chat-service/src/main/java/com/example/chatservice/util/Message.java
@@ -1,0 +1,13 @@
+package com.example.chatservice.util;
+
+public class Message {
+    public static final String CHATROOM_SAVE_SUCCESS_MESSAGE = "채팅방 정보 저장 성공";
+    public static final String CHATROOM_SAVE_FAIL_MESSAGE = "채팅방 정보 저장 실패";
+    public static final String CHATROOM_FIND_SUCCESS_MESSAGE = "채팅방 조회 성공";
+    public static final String CHATROOM_FIND_FAIL_MESSAGE = "채팅방 조회 실패";
+
+    public static final String MESSAGE_FIND_SUCCESS_MESSAGE = "메시지 조회 성공";
+    public static final String MESSAGE_FIND_FAIL_MESSAGE = "메시지 조회 실패";
+
+
+}
