+++ chat-service/src/test/java/com/example/chatservice/ChatApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.chatservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ChatApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
