--- chat/src/main/java/com/example/chat/ChatApplication.java
@@ -1,19 +1,0 @@
-package com.example.chat;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-
-@EnableCaching
-@EnableFeignClients
-@EnableEurekaClient
-@SpringBootApplication
-public class ChatApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ChatApplication.class, args);
-	}
-
-}
