--- chat/src/main/java/com/example/chat/ChatService.java
@@ -1,14 +1,0 @@
-package com.example.chat;
-
-import org.springframework.stereotype.Service;
-
-import java.util.Map;
-
-@Service
-public class ChatService {
-
-    // 채팅방 생성 후 저장
-//    public Map<String, Object> makeChatRoom() {}
-
-    // 채팅 내역 가져오기 -> 채팅방 id 통해 기록 가져오기
-}
