--- chat/src/main/java/com/example/chat/repository/ChatRoomRepository.java
@@ -1,8 +1,0 @@
-package com.example.chat.repository;
-
-import com.example.chat.model.entity.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
-}
