--- chat/src/main/java/com/example/chat/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.chat.repository;
-
-import com.example.chat.model.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, String> {
-    Optional<User> findByNickname(String nickname);
-}
