--- chat/src/test/java/com/example/chat/ChatApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.chat;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ChatApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
